create move renderPreviousExp and renderRelatedEd into their own components PreviousExp and RelatedEducation, as children of education component
